requirements.txt
django>=6.0
pillow>=12.0
requests>=2.32
